requirements.txt
alembic==1.4.3
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-Alembic==2.0.1
Flask-Cors==3.0.9
Flask-JWT-Extended==4.1.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
GeoAlchemy2==0.9.0
importlib-metadata==4.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pep517==0.10.0
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
typing-extensions==3.10.0.0
Werkzeug==1.0.1
zipp==3.4.1
